requirements.txt
clipboard==0.0.4
lxml==4.5.0
pyperclip==1.8.0
python-docx==0.8.10
